Refactor planner
