+++ src/analysis/sunmmio_tile_loop_fusion_planner.cc
@@ -7,134 +7,202 @@
 
 #include "sunmmio_tile_loop_fusion_planner_internal.h"
 
-#include <tvm/ffi/reflection/registry.h>
+#include <algorithm>
+#include <sstream>
+#include <unordered_map>
 
 namespace tvm {
 namespace tl {
 
-using namespace detail;
+namespace {
 
-Map<String, ffi::Any> DebugSunmmioTileLoopFusionPhase1Guardrails() {
-  auto make_resident = [](int64_t payload_bytes, int64_t instance_count) {
-    ResidentValueState resident;
-    resident.kind = ResidentValueKind::kDefinition;
-    resident.origin_region_local_index = 3;
-    resident.buffer_region_id = 7;
-    resident.buffer_name = "debug_buffer";
-    resident.home_depth = 2;
-    resident.payload_bytes = payload_bytes;
-    resident.instance_count = instance_count;
-    return resident;
+std::string BufferRegionKey(const tir::BufferRegion &region) {
+  std::ostringstream os;
+  os << region->buffer->name << '|';
+  for (const Range &range : region->region) {
+    os << PrimExprToString(range->min) << ':' << PrimExprToString(range->extent)
+       << '|';
+  }
+  return os.str();
+}
+
+int64_t ComputeExecutionPrefixInstanceCount(const Array<PrimExpr> &extents,
+                                            int depth) {
+  if (depth <= 0) {
+    return 1;
+  }
+  int64_t count = 1;
+  int limit = std::min(depth, static_cast<int>(extents.size()));
+  for (int i = 0; i < limit; ++i) {
+    const auto *imm = extents[i].as<tir::IntImmNode>();
+    if (imm == nullptr) {
+      return 1;
+    }
+    count = SaturatingMulSunmmioTileLoopFusionPlannerCost(count, imm->value);
+  }
+  return count;
+}
+
+int64_t
+ComputeRawEdgeInstanceCount(const Array<PrimExpr> &src_execution_loop_extents,
+                            const Array<PrimExpr> &dst_execution_loop_extents,
+                            int rho) {
+  return std::max(
+      ComputeExecutionPrefixInstanceCount(src_execution_loop_extents, rho),
+      ComputeExecutionPrefixInstanceCount(dst_execution_loop_extents, rho));
+}
+
+planner_internal::WindowPlannerInput
+BuildWindowPlannerInput(const SunmmioTileLoopFusionWindowProblem &problem) {
+  planner_internal::WindowPlannerInput input;
+  input.problem = &problem;
+  int num_regions = static_cast<int>(problem.regions.size());
+  input.regions.reserve(num_regions);
+  input.incoming_edges_by_dst.resize(num_regions);
+  input.outgoing_edges_by_src.resize(num_regions);
+  input.predecessor_masks.assign(num_regions,
+                                 planner_internal::DynamicBitset(num_regions));
+  input.earlier_source_masks.assign(
+      num_regions, planner_internal::DynamicBitset(num_regions));
+
+  std::unordered_map<int, int> local_index_by_global;
+  std::unordered_map<std::string, int> region_id_by_key;
+
+  auto get_region_id = [&](const tir::BufferRegion &region) {
+    std::string key = BufferRegionKey(region);
+    auto it = region_id_by_key.find(key);
+    if (it != region_id_by_key.end()) {
+      return it->second;
+    }
+    int id = static_cast<int>(region_id_by_key.size());
+    region_id_by_key.emplace(key, id);
+    return id;
   };
 
-  OpenScopeFrame identical_frame;
-  identical_frame.residents = {make_resident(64, 8), make_resident(64, 8)};
-  SortAndDedupeResidents(&identical_frame);
+  for (int local_index = 0; local_index < num_regions; ++local_index) {
+    const TileScopeRegion &region = problem.regions[local_index];
+    int global_region_index = region.global_region_index;
+    local_index_by_global[global_region_index] = local_index;
 
-  OpenScopeFrame payload_distinct_frame;
-  payload_distinct_frame.residents = {make_resident(64, 8),
-                                      make_resident(128, 8)};
-  SortAndDedupeResidents(&payload_distinct_frame);
+    const NormalizedTileScopeRegion &normalized =
+        problem.normalized_regions[local_index];
 
-  OpenScopeFrame instance_distinct_frame;
-  instance_distinct_frame.residents = {make_resident(64, 8),
-                                       make_resident(64, 16)};
-  SortAndDedupeResidents(&instance_distinct_frame);
+    planner_internal::WindowPlannerRegionInfo info;
+    info.global_region_index = global_region_index;
 
-  int64_t limit =
-      MakeInfiniteSunmmioTileLoopFusionPlannerScore().write_cut_cost;
-  Map<String, ffi::Any> summary;
-  summary.Set("planner_cost_limit", Integer(limit));
-  summary.Set(
-      "saturating_add_overflow",
-      Integer(SaturatingAddSunmmioTileLoopFusionPlannerCost(limit - 1, 42)));
-  summary.Set(
-      "saturating_mul_overflow",
-      Integer(SaturatingMulSunmmioTileLoopFusionPlannerCost(limit / 2 + 1, 3)));
-  summary.Set("resident_dedupe_identical_count",
-              Integer(static_cast<int64_t>(identical_frame.residents.size())));
-  summary.Set(
-      "resident_dedupe_payload_distinct_count",
-      Integer(static_cast<int64_t>(payload_distinct_frame.residents.size())));
-  summary.Set(
-      "resident_dedupe_instance_distinct_count",
-      Integer(static_cast<int64_t>(instance_distinct_frame.residents.size())));
-  return summary;
-}
+    for (const NormalizedBufferAccess &use_access : normalized.use_in) {
+      info.use_in.push_back(
+          {get_region_id(use_access.region),
+           static_cast<std::string>(use_access.region->buffer->name),
+           use_access.home_depth, use_access.payload_bytes,
+           ComputeExecutionPrefixInstanceCount(region.execution_loop_extents,
+                                               use_access.home_depth)});
+    }
+    for (const NormalizedBufferAccess &def_access : normalized.def_out) {
+      info.def_out.push_back(
+          {get_region_id(def_access.region),
+           static_cast<std::string>(def_access.region->buffer->name),
+           def_access.home_depth, def_access.payload_bytes,
+           ComputeExecutionPrefixInstanceCount(region.execution_loop_extents,
+                                               def_access.home_depth)});
+    }
 
-bool DebugSunmmioTileLoopFusionCheckBitsetBounds(int num_bits, int index) {
-  DynamicBitset bitset(num_bits);
-  SetBit(&bitset, index);
-  return TestBit(bitset, index);
-}
-
-Map<String, ffi::Any> DebugSunmmioTileLoopFusionRawCoverageAccounting() {
-  WindowPlannerInput input;
-  input.region_indices = {0, 1, 2};
-  input.regions.resize(3);
-  input.edges.resize(2);
-  input.incoming_edges_by_dst.resize(3);
-  input.outgoing_edges_by_src.resize(3);
-  input.predecessor_masks.assign(3, DynamicBitset(3));
-  input.earlier_source_masks.assign(3, DynamicBitset(3));
-
-  for (int i = 0; i < 3; ++i) {
-    input.regions[i].global_region_index = i;
-    input.regions[i].source_order_index = i;
-    input.regions[i].logical_execution_axes = {"i"};
-    input.regions[i].execution_loop_extents = {IntImm(DataType::Int(32), 1)};
+    input.regions.push_back(std::move(info));
+    for (int earlier = 0; earlier < local_index; ++earlier) {
+      planner_internal::SetBit(&input.earlier_source_masks[local_index],
+                               earlier);
+    }
   }
 
-  input.regions[1].use_in.push_back(
-      {10, "debug_buffer", tir::BufferRegion(), 1, 64, 1});
-  input.regions[1].use_in.push_back(
-      {11, "debug_buffer", tir::BufferRegion(), 1, 80, 1});
-  input.regions[2].use_in.push_back(
-      {10, "debug_buffer", tir::BufferRegion(), 1, 64, 1});
+  input.edges.reserve(problem.graph.edges.size());
+  for (const TileScopeDependenceEdge &edge : problem.graph.edges) {
+    auto src_it = local_index_by_global.find(edge.src_region_index);
+    auto dst_it = local_index_by_global.find(edge.dst_region_index);
+    if (src_it == local_index_by_global.end() ||
+        dst_it == local_index_by_global.end()) {
+      continue;
+    }
 
-  WindowPlannerEdgeInfo first_edge;
-  first_edge.src_local_index = 0;
-  first_edge.dst_local_index = 1;
-  first_edge.kind = TileScopeDependenceKind::kRAW;
-  first_edge.buffer_region_id = 10;
-  first_edge.buffer_name = "debug_buffer";
-  first_edge.rho = 1;
-  first_edge.weight = 64;
-  first_edge.instance_count = 1;
-  first_edge.covered_use_indices = {0};
-  input.edges[0] = first_edge;
+    planner_internal::WindowPlannerEdgeInfo planner_edge;
+    planner_edge.src_local_index = src_it->second;
+    planner_edge.dst_local_index = dst_it->second;
+    planner_edge.kind = edge.kind;
+    planner_edge.rho = edge.rho;
+    planner_edge.weight = edge.weight_bytes;
+    planner_edge.instance_count = ComputeRawEdgeInstanceCount(
+        problem.regions[planner_edge.src_local_index].execution_loop_extents,
+        problem.regions[planner_edge.dst_local_index].execution_loop_extents,
+        planner_edge.rho);
 
-  WindowPlannerEdgeInfo second_edge = first_edge;
-  second_edge.dst_local_index = 2;
-  input.edges[1] = second_edge;
+    if (edge.kind == TileScopeDependenceKind::kRAW) {
+      const auto &uses = input.regions[planner_edge.dst_local_index].use_in;
+      ICHECK_GE(edge.dst_access_index, 0);
+      ICHECK_LT(edge.dst_access_index, static_cast<int>(uses.size()));
+      const planner_internal::PlannerBufferValueInfo &covered_use =
+          uses[edge.dst_access_index];
+      planner_edge.buffer_region_id = covered_use.buffer_region_id;
+      planner_edge.buffer_name = covered_use.buffer_name;
+      planner_edge.covered_use_index = edge.dst_access_index;
+    } else if (edge.kind == TileScopeDependenceKind::kWAR) {
+      const auto &uses = input.regions[planner_edge.src_local_index].use_in;
+      ICHECK_GE(edge.src_access_index, 0);
+      ICHECK_LT(edge.src_access_index, static_cast<int>(uses.size()));
+      const planner_internal::PlannerBufferValueInfo &source_use =
+          uses[edge.src_access_index];
+      planner_edge.buffer_region_id = source_use.buffer_region_id;
+      planner_edge.buffer_name = source_use.buffer_name;
+    } else {
+      const auto &defs = input.regions[planner_edge.dst_local_index].def_out;
+      ICHECK_GE(edge.dst_access_index, 0);
+      ICHECK_LT(edge.dst_access_index, static_cast<int>(defs.size()));
+      const planner_internal::PlannerBufferValueInfo &dst_def =
+          defs[edge.dst_access_index];
+      planner_edge.buffer_region_id = dst_def.buffer_region_id;
+      planner_edge.buffer_name = dst_def.buffer_name;
+    }
 
-  input.incoming_edges_by_dst[1].push_back(0);
-  input.incoming_edges_by_dst[2].push_back(1);
-  input.outgoing_edges_by_src[0].push_back(0);
-  input.outgoing_edges_by_src[0].push_back(1);
-  SetBit(&input.predecessor_masks[1], 0);
-  SetBit(&input.predecessor_masks[2], 0);
-  SetBit(&input.earlier_source_masks[1], 0);
-  SetBit(&input.earlier_source_masks[2], 0);
-  SetBit(&input.earlier_source_masks[2], 1);
+    int edge_index = static_cast<int>(input.edges.size());
+    input.edges.push_back(planner_edge);
+    input.incoming_edges_by_dst[planner_edge.dst_local_index].push_back(
+        edge_index);
+    input.outgoing_edges_by_src[planner_edge.src_local_index].push_back(
+        edge_index);
+    planner_internal::SetBit(
+        &input.predecessor_masks[planner_edge.dst_local_index],
+        planner_edge.src_local_index);
 
-  DynamicBitset raw_consumers(3);
-  SetBit(&raw_consumers, 1);
-  SetBit(&raw_consumers, 2);
-  input.raw_consumer_masks_by_key.emplace(RawConsumerKey(0, 10), raw_consumers);
+    if (planner_edge.kind == TileScopeDependenceKind::kRAW) {
+      planner_internal::RawConsumerKey key{planner_edge.src_local_index,
+                                           planner_edge.buffer_region_id};
+      auto it = input.raw_consumer_masks_by_key.find(key);
+      if (it == input.raw_consumer_masks_by_key.end()) {
+        it = input.raw_consumer_masks_by_key
+                 .emplace(key, planner_internal::DynamicBitset(num_regions))
+                 .first;
+      }
+      planner_internal::SetBit(&it->second, planner_edge.dst_local_index);
+    }
+  }
 
-  PlannerState state{DynamicBitset(3), {}};
-  TransitionResult first = ApplyAction(input, state, 1, 0, 1);
-  TransitionResult second = ApplyAction(input, first.next_state, 2, 1, 1);
+  for (int local_index = 0; local_index < num_regions; ++local_index) {
+    for (const planner_internal::PlannerBufferValueInfo &use_info :
+         input.regions[local_index].use_in) {
+      auto it = input.read_consumer_masks_by_region_id.find(
+          use_info.buffer_region_id);
+      if (it == input.read_consumer_masks_by_region_id.end()) {
+        it = input.read_consumer_masks_by_region_id
+                 .emplace(use_info.buffer_region_id,
+                          planner_internal::DynamicBitset(num_regions))
+                 .first;
+      }
+      planner_internal::SetBit(&it->second, local_index);
+    }
+  }
 
-  Map<String, ffi::Any> summary;
-  summary.Set("first_write_cut_cost", Integer(first.delta.write_cut_cost));
-  summary.Set("first_shared_read_cost", Integer(first.delta.shared_read_cost));
-  summary.Set("second_write_cut_cost", Integer(second.delta.write_cut_cost));
-  summary.Set("second_shared_read_cost",
-              Integer(second.delta.shared_read_cost));
-  return summary;
+  return input;
 }
+
+} // namespace
 
 std::vector<SunmmioTileLoopFusionWindowPlan>
 PlanSunmmioTileLoopFusionWindowProblems(
@@ -143,45 +211,41 @@ PlanSunmmioTileLoopFusionWindowProblems(
   plans.reserve(problems.size());
 
   for (const SunmmioTileLoopFusionWindowProblem &problem : problems) {
-    WindowPlannerInput input = BuildWindowPlannerInput(problem);
+    planner_internal::WindowPlannerInput input =
+        BuildWindowPlannerInput(problem);
 
-    MemoResult best;
-    if (static_cast<int>(input.regions.size()) > kMaxExactPlannerRegions) {
-      best = BuildSourceOrderFallbackPlan(input);
+    planner_internal::MemoResult best;
+    if (static_cast<int>(input.regions.size()) >
+        planner_internal::kMaxExactPlannerRegions) {
+      best = planner_internal::BuildSourceOrderFallbackPlan(input);
     } else {
-      PlannerState initial_state{
-          DynamicBitset(static_cast<int>(input.regions.size())), {}};
-      PlannerSearchContext context;
-      best = SolveWindowPlan(input, initial_state, &context);
+      planner_internal::PlannerState initial_state{
+          planner_internal::DynamicBitset(
+              static_cast<int>(input.regions.size())),
+          {}};
+      planner_internal::PlannerSearchContext context;
+      best = planner_internal::SolveWindowPlan(input, initial_state, &context);
       if (context.exhausted) {
-        best = BuildSourceOrderFallbackPlan(input);
+        best = planner_internal::BuildSourceOrderFallbackPlan(input);
       }
     }
 
     SunmmioTileLoopFusionWindowPlan summary;
-    summary.region_indices = input.region_indices;
+    summary.region_indices.reserve(input.regions.size());
+    for (const planner_internal::WindowPlannerRegionInfo &region :
+         input.regions) {
+      summary.region_indices.push_back(region.global_region_index);
+    }
     summary.score = best.score;
     summary.actions = best.actions;
     for (const SunmmioTileLoopFusionPlannerAction &action : best.actions) {
       summary.order.push_back(action.region_index);
     }
-    summary.tree = BuildPlanTree(best.actions);
+    summary.tree = planner_internal::BuildPlanTree(best.actions);
     plans.push_back(std::move(summary));
   }
 
   return plans;
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  tvm::ffi::reflection::GlobalDef().def(
-      "tl.analysis.DebugSunmmioTileLoopFusionPhase1Guardrails",
-      DebugSunmmioTileLoopFusionPhase1Guardrails);
-  tvm::ffi::reflection::GlobalDef().def(
-      "tl.analysis.DebugSunmmioTileLoopFusionCheckBitsetBounds",
-      DebugSunmmioTileLoopFusionCheckBitsetBounds);
-  tvm::ffi::reflection::GlobalDef().def(
-      "tl.analysis.DebugSunmmioTileLoopFusionRawCoverageAccounting",
-      DebugSunmmioTileLoopFusionRawCoverageAccounting);
 }
 
 } // namespace tl
